Add code to process VOC files. Add code to play VOC files in direct play mode on SB16 (might work on real hardware but doesn't work in Qemu). Much of this code is temporary and will be rewritten for DMA. Improve PIT code a little and add PIT_Set_Interval(). Don't let exception handlers return if interrupted. Add memcmp() function.

+++ MyOS_1/Graphics/Display_HAL.c
@@ -22,7 +22,7 @@ void GraphicsBlit(unsigned int x, unsigned int y, PIXEL_32BIT *image, uint32_t w
     PIXEL_32BIT *currentPixel = (PIXEL_32BIT *)((uint32_t)linearFrameBuffer + framebufferOffset);
 
     // for every row of pixels
-    for (int h = 0; h < height; ++h)
+    for (size_t h = 0; h < height; ++h)
     {
         // copy the current row
         memcpy((void*)currentPixel, (const void *)image, sizeof(PIXEL_32BIT) * width);
